chore: simplifies relatedHelpList comp.
