+++ src/app/(pages)/docs/[topic]/[doc]/client_page.tsx
@@ -8,7 +8,7 @@ import { MDXRemote } from 'next-mdx-remote'
 import { Button } from '@components/Button'
 import { JumplistProvider } from '@components/Jumplist'
 import components from '@components/MDX/components'
-import { RelatedDiscordGitList } from '@components/RelatedDiscordGitList'
+import { RelatedHelpList } from '@components/RelatedHelpList'
 import TableOfContents from '@components/TableOfContents'
 import slugify from '@root/utilities/slugify'
 import { Doc, NextDoc } from '../../types'
@@ -52,7 +52,7 @@ export const RenderDoc: React.FC<Props> = ({ doc, next }) => {
           <div className={classes.mdx}>
             <MDXRemote {...content} components={components} />
           </div>
-          <RelatedDiscordGitList />
+          <RelatedHelpList />
           {next && (
             <Link
               className={classes.next}
