--- src/components/RelatedDiscordGitList/index.module.scss
@@ -1,128 +1,0 @@
-@use '@scss/common' as *;
-
-.relatedDiscordGitList {
-  position: relative;
-
-  :global(.rah-static) {
-    position: relative;
-
-    &::after {
-      content: "";
-      position: absolute;
-      width: 100%;
-      height: 100%;
-      bottom: 0;
-      pointer-events: none;
-      z-index: 2;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      top: 0;
-      opacity: 1;
-      will-change: opacity;
-      transition: opacity var(--default-trans-time) linear;
-    }
-  }
-
-  :global(.rah-static--height-auto) {
-    &::after {
-      opacity: 0;
-    }
-  }
-
-  .collapsibleToggler {
-    all: unset;
-    cursor: pointer;
-    box-sizing: border-box;
-    display: inline-flex;
-    position: relative;
-    height: calc(var(--base) * 2.5); // match input height
-
-    &:focus-visible {
-      @include outline;
-    }
-
-    &:disabled {
-      cursor: not-allowed;
-      opacity: .25;
-    }
-
-    display: inline-flex;
-    width: 100%;
-    padding: unset;
-    padding-top: calc(var(--base) * 0.75);
-  }
-  
-  .togglerOpen {
-    position: sticky;
-    bottom: 0;
-    text-align: center;
-    width: 100%;
-    z-index: 1;
-  }
-
-  .titleWrapper {
-    padding-top: calc(var(--base) * 2);
-    margin-bottom: var(--base);
-  }
-  
-  .title {
-    margin: 0;
-  }
-
-  .togglerContent {
-    position: relative;
-    display: flex;
-    align-items: center;
-    width: 100%;
-  }
-  
-  .inputArrow {
-    z-index: 1;
-    color: var(--theme-text);
-    transform: translateY(-50%) rotate(135deg);
-    height: calc(var(--base) * 0.5);
-    width: calc(var(--base) * 0.5);
-  }
-  
-  .isOpenArrow {
-    transform: translateY(-50%) rotate(315deg);
-  }
-  
-  .togglerLabel {
-    @include label;
-    width: 100%;
-    flex-grow: 1;
-    flex-shrink: 1;
-    display: flex;
-    align-items: center;
-  }
-}
-
-:global([data-theme="dark"]) {
-  .relatedDiscordGitList {
-    :global(.rah-static) {
-      &::after {
-        background: linear-gradient(0deg, rgba(20, 20, 20, 1) 25%, rgba(20, 20, 20, 0));
-      }
-    }
-
-    .togglerOpen {
-      background-color: var(--color-base-900);
-    }
-  }
-}
-
-:global([data-theme="light"]) {
-  .relatedDiscordGitList {
-    :global(.rah-static) {
-      &::after {
-        background: linear-gradient(0deg, rgba(255, 255, 255, .98) 25%, rgba(255, 255, 255, 0));
-      }
-    }
-
-    .togglerOpen {
-      background-color: var(--color-base-0);
-    }
-  }
-}
