+++ src/components/RelatedHelpList/index.module.scss
@@ -1,0 +1,40 @@
+@use '@scss/common' as *;
+
+.relatedHelpList {
+  position: relative;
+
+  .titleWrapper {
+    padding-top: calc(var(--base) * 1.5);
+    margin-bottom: var(--base);
+  }
+  
+  .title {
+    margin: 0;
+  }
+
+  ul {
+    display: flex;
+    gap: calc(var(--base) * 0.25);
+    flex-direction: column;
+    list-style: none;
+    padding-left: 0;
+    margin-bottom: calc(var(--base) * 1.5);
+  }
+
+  li {
+    position: relative;
+    padding-left: calc(var(--base) * 1.5);
+  }
+
+  .itemMarker {
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 20px;
+    margin: 7px 0 0 0;
+
+    @include small-break {
+      margin: 3px 0 0 0;
+    }
+  }
+}
